Updated the notification controllers and added the create notification function to goal controllers

+++ controllers/notificationController.js
@@ -24,10 +24,19 @@ const notificationStructure = {
     expiredGoal: [
         'We failed to reach this goal.',
         'Unfortunately, you have been unable to achieve this goal within the set timeframe.'
+    ],
+    deleteGoal: [
+        // eslint-disable-next-line quotes
+        "A goal you're assigned to has been deleted.",
+        'We will no longer be working towards this goal.'
+    ],
+    unassignGoal: [
+        'You have been unassigned from this goal',
+        'You will no longer get updates for this goal.'
     ]
 };
 
-exports.createNotification = async (userId, orgId, goalName, funcName) => {
+exports.createNotification = async (req, res, userId, orgId, goalName, funcName) => {
     const notification = {
         user_id: userId,
         org_id: orgId,
@@ -37,11 +46,21 @@ exports.createNotification = async (userId, orgId, goalName, funcName) => {
         description: notificationStructure[funcName][1],
         createdAt: Date.now()
     };
-    await notificationSchema.validateAsync(notification);
-    await insertOne('notifications', notification, orgId);
+
+    try {
+        await notificationSchema.validateAsync(notification);
+    } catch (error) {
+        return res.status(400).json(error.details)
+    }
+
+    try {
+        await insertOne('notifications', notification, orgId);
+    } catch (error) {
+        return res.status(500).json(error.details)
+    }
 };
 
-exports.getUserNotifications = catchAsync(async (req, res, next) => {
+exports.getUserNotifications = async (req, res, next) => {
     const {
         org_id: orgId,
         user_id: userId
@@ -60,7 +79,7 @@ exports.getUserNotifications = catchAsync(async (req, res, next) => {
     }
     try {
         // Search for all Goals
-        const goals = await find('notifications', {
+        const notifications = await find('notifications', {
             org_id: orgId,
             user_id: userId
         }, orgId);
@@ -69,16 +88,17 @@ exports.getUserNotifications = catchAsync(async (req, res, next) => {
         res.status(200).json({
             status: 200,
             message: 'success',
-            data: goals.data.data
+            data: notifications.data.data
         })
     } catch (error) {
-        res.status(500).json({
-            status: 500,
-            message: 'There was an error processing this request.'
+          res.status(200).json({
+            status: 200,
+            // eslint-disable-next-line quotes
+            message: "You don't have any notifications."
         })
     }
 
-});
+};
 
 exports.updateNotification = catchAsync(async (req, res, next) => {
     const {
@@ -107,10 +127,14 @@ exports.updateNotification = catchAsync(async (req, res, next) => {
         status: 'read'
     }
     try {
-        await updateOne('notifications', update, {}, orgId, notificationId)
+        await updateOne('notifications', update, { _id: notificationId }, orgId, notificationId)
         return res.status(200).json('Notification read.')
     } catch (error) {
-        return next(new AppError('An error has eoccured while processing this request', 500));
+        res.status(500).json({
+            status: 500,
+            // eslint-disable-next-line quotes
+            message: "Unable to update this notification"
+        })
     }
 });
 
@@ -118,35 +142,6 @@ exports.updateNotifications = catchAsync(async (req, res, next) => {
     const {
         org_id: orgId,
         user_id: userId
-    } = req.query;
-
-    // Check for user_id and organization_id
-    if (!orgId) {
-        return next(new AppError('organization_id is required', 403));
-    }
-    if (!userId) {
-        return next(new AppError('user_id is required', 403));
-    }
-
-    const update = {
-        status: 'read'
-    }
-    try {
-        await updateMany('notifications', update, {
-            org_id: orgId,
-            user_id: userId
-        }, orgId)
-        return res.status(200).json('All notifications read.')
-    } catch (error) {
-        return next(new AppError('An error has eoccured while processing this request', 500));
-    }
-});
-
-exports.deleteNotification = catchAsync(async (req, res, next) => {
-    const {
-        org_id: orgId,
-        user_id: userId,
-        notification_id: notificationId
     } = req.query;
 
     // Check for org_id and user_id
@@ -160,15 +155,59 @@ exports.deleteNotification = catchAsync(async (req, res, next) => {
             error: 'user_id is required'
         })
     }
+
+    const update = {
+        status: 'read'
+    }
+    try {
+        await updateMany('notifications', update, {
+            org_id: orgId,
+            user_id: userId
+        }, orgId)
+        return res.status(200).json('All notifications read.')
+    } catch (error) {
+        res.status(500).json({
+            status: 500,
+            // eslint-disable-next-line quotes
+            message: "Unable to mark all notifications read."
+        })
+    }
+});
+
+exports.deleteNotification = catchAsync(async (req, res, next) => {
+    const {
+        org_id: orgId,
+        user_id: userId,
+        notification_id: notificationId
+    } = req.query;
+
+    // Check for org_id, user_id and notification_id
+    if (!orgId) {
+        return res.status(403).send({
+            error: 'org_id is required'
+        })
+    }
+
+    if (!userId) {
+        return res.status(403).send({
+            error: 'user_id is required'
+        })
+    }
+
     if (!notificationId) {
         return res.status(403).send({
             error: 'notification_id is required'
         })
     }
+    
     try {
         await deleteOne('notifications', orgId, notificationId)
         return res.status(200).json('Notification deleted.')
     } catch (error) {
-        return next(new AppError('An error has eoccured while processing this request', 500));
+        res.status(500).json({
+            status: 500,
+            // eslint-disable-next-line quotes
+            message: "Unable to delete this notification"
+        })
     }
 })
